feat: refactor processor code
Consolidate actions into fewer files: one file per entity.

Get rid of the common interface for actions as it serves no real
purpose.

Type-cast objects only once after unmarshalling (rather than once per
action).

Consolidate actions' interfaces to accept the right type for such
action:

 - PodAction accepts v1.Pod
 - NodeAction accepts v1.Node,
 - ... actions for new entity types will do the same
